Add airbnb eslint standars and adapt the code
