+++ package.json
@@ -41,6 +41,7 @@
   "devDependencies": {
     "@testing-library/react": "^13.4.0",
     "eslint": "^8.27.0",
+    "eslint-config-airbnb": "^19.0.4",
     "eslint-plugin-react": "^7.31.10"
   }
 }
