+++ src/components/Login/Login.module.css
@@ -1,18 +1,18 @@
 .container {
-    display: flex;
-    position: absolute;
-    justify-content: center;
     align-items: center;
-    flex-direction: column;
-    width: 100%;
-    height: 100%;
     background: linear-gradient(1turn, #f5efe6, 60%, #aebdca);
+    display: flex;
+    flex-direction: column;
+    height: 100%;
+    justify-content: center;
+    position: absolute;
+    width: 100%;
   }
 
   .contLogo {
+    align-items: center;
     display: flex;
     flex-direction: column;
-    align-items: center;
     width: 50%;
   }
   
@@ -21,41 +21,41 @@
   }
   
   .contform {
-    display: flex;
-    justify-content: center;
     align-items: center;
+    display: flex;
     flex-direction: column;
-    width: 40%;
     height: 40%;
+    justify-content: center;
     margin: 2%;
+    width: 40%;
   }
 
   .contform form {
-    width: 100%;
-    display: flex;
     align-items: center;
-    justify-content: center;
-    flex-direction: column;
     border-radius: 12px;
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    width: 100%;
   }
   
   .campo {
+    align-items: center;
     display: flex;
     flex-direction: column;
-    align-items: center;
-    width: 90%;
     margin-top: 5%;
+    width: 90%;
   }
   
   .campo label {
     font-size: 20px;
-    height: 30px;
     font-weight: bold;
+    height: 30px;
   }
   
   .btnUser {
-    border: 1px solid #393939;
     border-radius: 5px 5px 5px 5px;
+    border: 1px solid #393939;
     color: #393939;
     font-size: 20px;
     padding: 5px;
@@ -63,8 +63,8 @@
   }
 
   .btnPass {
-    border: 1px solid #393939;
     border-radius: 5px 5px 5px 5px;
+    border: 1px solid #393939;
     color: #393939;
     font-size: 20px;
     padding: 5px;
@@ -73,15 +73,15 @@
   
   .btnSend {
     background: #aebdca;
+    border-radius: 5px 5px 5px 5px;
     border: 1px solid #000;
     color: #000;
-    border-radius: 5px 5px 5px 5px;
+    cursor: pointer;
     font-size: 12px;
-    padding: 5px;
-    width: 90%;
     font-size: 20px;
     margin: 5%;
-    cursor: pointer;
+    padding: 5px;
+    width: 90%;
   }
 
   .btnSend:hover {
@@ -102,8 +102,8 @@
     }
 
     .campo {
-      width: 100%;
       margin: 5px;
+      width: 100%;
     }
 
     .imgLogo {
@@ -117,22 +117,22 @@
   
   @media screen and (max-width: 1200px) {
     .contform form {
-      width: 100%;
       margin-top: 25px;
+      width: 100%;
     }
 
     .campo {
-      width: 100%;
       margin: 5px;
+      width: 100%;
     }
 
     .contLogo {
-      display: flex;
       align-items: center;
+      display: flex;
       justify-content: center;
-      width: 100%;
-      margin-top: 10%;
       margin-bottom: 5%;
+      margin-top: 10%;
+      width: 100%;
     }
   }
