+++ src/reportWebVitals.js
@@ -1,4 +1,4 @@
-const reportWebVitals = onPerfEntry => {
+const reportWebVitals = (onPerfEntry) => {
 	if (onPerfEntry && onPerfEntry instanceof Function) {
 		import('web-vitals').then(({ getCLS, getFID, getFCP, getLCP, getTTFB }) => {
 			getCLS(onPerfEntry);
